# frontend/openapi/exam.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "天机学堂 - 考试中心接口文档",
    "description": "该服务提供题目管理、考试管理、数据统计等功能",
    "contact": {
      "name": "传智教育·研究院",
      "url": "http://www.itcast.cn/",
      "email": "zhanghuyi@itcast.cn"
    },
    "version": "v1.0"
  },
  "paths": {
...
    "/api/v2/exam-attempts/{id}/draft": {
      "get": {
        "tags": [
          "exam-draft-controller"
        ],
        "operationId": "read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/View"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "exam-draft-controller"
        ],
        "operationId": "save",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Save"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/View"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/teacher/exam-attempts/{id}/grades": {
      "post": {
        "tags": [
          "exam-workflow-controller"
        ],
        "operationId": "grade",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/View"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/exam-papers/{id}/attempts": {
      "post": {
        "tags": [
          "exam-workflow-controller"
        ],
        "operationId": "start",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/View"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/exam-attempts/{id}/submit": {
      "post": {
        "tags": [
          "exam-workflow-controller"
        ],
        "operationId": "submit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/View"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/admin/exam-papers": {
      "post": {
        "tags": [
          "exam-workflow-controller"
        ],
        "operationId": "publish",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/View"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/teacher/exam-attempts/{id}": {
      "get": {
        "tags": [
          "exam-workflow-controller"
        ],
        "operationId": "grading",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/operations/exam/{id}": {
      "get": {
        "tags": [
          "operation-controller"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/View"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/exam-papers": {
      "get": {
        "tags": [
          "exam-workflow-controller"
        ],
        "operationId": "papers",
        "parameters": [
          {
            "name": "courseId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/exam-attempts/{id}": {
      "get": {
        "tags": [
          "exam-workflow-controller"
        ],
        "operationId": "attempt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/admin/consumer-failures/exam": {
      "get": {
        "tags": [
          "consumer-failure-controller"
        ],
        "operationId": "list_1",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Save": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "answers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "View": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "answers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "Command": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "paperId": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "attemptId": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "lessonId": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "answers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "questionId": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "version": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "feedback": {
            "type": "string"
          },
          "actorRole": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          }
        }
      },
      "PublishRequest": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "sectionId": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "questionIds": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          },
          "passPercent": {
            "type": "integer",
            "format": "int32"
          },
          "graders": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            }
          }
        }
      },
...
    }
  }
}